refactor: immutable view plus builder
